execution_server/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
black==22.12.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.4
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2022.7
redis==4.5.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
django-cors-headers==3.13.0
